Null-check WebContents before calling into it (#378)
This protects against the case when
`blpwtk2::WebViewImpl::GetDefaultActivationWindow` is called after
`WebViewImpl::destroy` but before `WebViewImpl`'s destructor. Since the
`destroy` function schedules an asynchronous task to delete itself,
there is a small window of time for the call to
`GetDefaultActivationWindow` to sneak in.

The function `views::DesktopNativeWidgetAura::HandleActivationChanged`
makes a call to `GetDefaultActivationWindow` only if the `WebViewImpl`
instance is available. If it's not available, it assumes a return value
of `nullptr`. By adding the null-check for `d_webContents` in
`GetDefaultActivationWindow`, we are explicitly returning a `nullptr` if
the associated `content::WebContents` is already released. This would
have the same effect as if `blpwtk2::NativeViewWidget`'s delegate was
released by `WebViewImpl::destroy` and `HandleActivationChanged` thought
the `WebViewImpl` is no longer available.