[m98] Fix build for `ext/focus` branch (#402)
This change removes the dependency on `web_view_impl.h`.  Including this
header outside blink seems to cause a build error in one of the
transitive includes related to V8 double conversion.

Instead of using `blink::WebViewImpl` within `blpwtk2`, this change
makes `SetPageFocus` a virtual method so that it can be called via the
`blink::WebView` pointer.