Get d3dcompiler_47.dll from the distrib/win directory (#436)
In m93, chromium upstream updated the version of d3dcompiler_47.dll by
copying the new DLL to their WinSDK package:
https://source.chromium.org/chromium/chromium/src/+/e76661c8f279523537e0c4d99646083a604d9847

This would have required us to repackage the WinSDK as well.  However,
the way CEF deals with this is much simpler:
https://bitbucket.org/chromiumembedded/cef/commits/e63e6b110bde46e2858705e386fad53b325123c2

So the d3dcompiler_47.dll is just added to the source code repo (these
dlls are pretty small), and we change the `BUILD` step to copy the dlls
from the source code repo instead of copying it from the WinSDK.

This way, we don't need to hack up a new WinSDK package every time this
DLL is changed.

Note that the DLLs added in this commit are taken from the CEF branch
and are digitally signed by Microsoft.