Set drag image only if image contains a pixel buffer (#446)
This is an amendment to an older commit fcb95f27. In the previous
commit, we were checking for validity of the image after the creation of
the `gfx::ImageSkia`. In this changeset, we are moving up the check so
that the `gfx::ImageSkia` is not created if the source bitmap does not
contain any pixel data.