Call `setLogicalFocus(true)` at the end of `takeKeyboardFocus()` (#456)
This is a quick fix for the focus issue and is equivalent to the
embedder calling `setLogicalFocus(true)` immediately after calling
`takeKeyboardFocus()`.

This change avoids having to change all the places in embedder code
where `takeKeyboardFocus()` is called.

Eventually, we want to replace `takeKeyboardFocus()` and
`setLogicalFocus(bool)` with a single `setFocus(bool)` method.  At that
point, we can update all the embedders once and for all.