Fix potential deadlock when renderer establishes GPU channel (#465)
In the `RENDERER_MAIN` mode, the browser thread can potentially be
blocked by the renderer due to the window heirarchy (e.g. activating or
focusing the webview in the browser thread will send message to the
top-level hwnd in the renderer thread).

If the renderer is trying to establish the GPU channel at the same time,
it makes a sync request to the browser.  This would cause a deadlock.

This commit makes the renderer thread establish the GPU channel ahead of
time, before any potential hwnd interactions can take place.

Context: https://jira.prod.bloomberg.com/browse/ENG2KITFC-1256