[m98] Fix `-bb-lcd-background_color` functionality (#470)
Previously, the LCD background color was computed inside the
`CompositedLayerMapping` class and applied to the `GraphicsLayer`.
However, in `m98`, these classes are no longer used.

In the new code, layers are created using `PaintArtifactCompositor`.  A
`PaintArtifact` is a collection of `DisplayItem`s and `PaintChunk`s.
The decision to start a new layer or not is made based on the properties
of the `PaintChunk` (i.e. whether it has different transforms, clips,
effects; see `ScopedPaintChunkProperties`).

This commit adds a new `ScopedBbLcdBackgroundColor` class, which is used
to set the LCD background color of any `PaintChunk` created within its
scope.  Whenever we start using a new layer, the layer's default LCD
background color will be set to the first `PaintChunk`'s LCD background
color.

This commit also changes the way the `-bb-lcd-background-color` style is
computed when it is set to `auto`.  Previously, there was a notion of
the "compositing container", which was used to walk up the heirarchy at
paint time to determine the inherited background color.  Now we do the
following:

* Add a new field in `ComputedStyle` called
  `EffectiveBackgroundColorForAutoLcd`, which will contain the inherited
  background color (note: the `BackgroundColor` property itself is not
  inherited).
* Whenever the `BackgroundColor` property changes, we will compute the
  `EffectiveBackgroundColorForAutoLcd` using the new `BackgroundColor`
  and potentially the parent's `EffectiveBackgroundColorForAutoLcd`.
* If the `-bb-lcd-background-color` property is set to `auto`, then the
  `BbLcdBackgroundColor` field in the `ComputedStyle` will be set to the
  current `EffectiveBackgroundColorForAutoLcd`.

This commit removes the dead code in `CompositedLayerMapping` and
`GraphicsLayer` so that there will be no conflicts there in future
migrations (these classes have been removed upstream).

This commit also undoes some minor changes (`#include`, forward
declarations) in `skia` that are no longer needed.

Context: https://jira.prod.bloomberg.com/browse/ENG2KITFC-1247