Allow JS to use the new Async clipboard API
This commit adds a PermissionManager class, which implements the PermissionControllerDelegate
interface. It will be returned when BrowserContext.GetPermissionControllerDelegate() is called

This permission manager will allow clipboard access and deny other permission requests.
(Currently all permissions are denied because BrowserContext.GetPermissionControllerDelegate()
returns nullptr)

Test:
The tests in the bottom of https://web.dev/async-clipboard/