Rate limit GPU log messages (#502)
There are times when a flood of GPU related warning and error messages
can be generated. When these messages are relayed to the embedder, the
embedder can potentially flood log files as well.

This changeset applies a rate limiter to the messages based on the token
bucket algorithm. It allows a steady stream of 2 messages per second
and allows for a burst of 100 messages per second.

Test Plan
---------
1. Modify `kLogMessageLimit` to `1`, `token_size` to `1000` and `burst_limit` to `2`
2. Add a call to `observer_list_->Notify` in the `else` block (where we return early) with a custom message.
3. In `GpuChannelMessageFilter::OnReceiveMessage`, add a custom logger message
4. Check the log file generated by the browser process for the custom messages you added.