Release resource context after a successful load (#519)
Background and Motivation
-------------------------
The resource context is a sink that is created by the resource loader
bridge to let the embedder provide the response data for a particular
HTTP request.

This sink (resource context) is owned by blpwtk2 but it's kept alive
until the embedder is done with it. Once the embedder calls the
`finish` method, it no longer needs to use the context.

Change description
------------------
Once the `finish` method is called by the embedder, the context is
removed from the map that is used to keep it alive. This frees up the
memory that was allocated to hold the response data.