Generate a devkit header hash
The `blpwtk2_version` generated component used to have a unique symbol
for each blpwtk2 devkit version.  The intention behind it was to enforce
that embedders could only load blpwtk2.dll if they were built against
the same devkit version (the symbol would be missing if they were built
against any other version).

However, it looks like the mechanism that was used to force the linker
to pull in the symbol no longer works.  This is evident by looking at
the import table of the embedder binaries.

Starting in m98, we no longer embed the devkit version in the blpwtk2
exe/dll filenames (and we have also reverted back to using the standard
filenames for all the dlls like EGL and GLESv2).

In order to ensure that embedders will only load a `blpwtk2.dll` that is
ABI-compatible with that the embedder source files were built against,
we will generate a hash from all the public header files.

This hash will be embedded inside `blpwtk2.dll` and exposed to embedders
via the `getDevkitHeaderHash()` function.  When the embedder starts up,
it can verify that the hash provided by `blpwtk2.dll` matches the hash
that it sees in `BLPWTK2_DEVKIT_HEADER_HASH`.  This will ensure that it
will not load a `blpwtk2.dll` that is ABI-incompatible.

Note that if `blpwtk2` was built locally outside a git directory (or if
`git` is not available in the PATH), then an invalid hash will be
returned, which will ensure that embedders that were built against a
devkit will not work.

Context: https://jira.prod.bloomberg.com/browse/ENG2BBKIT-7259