Allow MSVC embedders to create a `v8::CppHeap`
The public function `v8::CppHeap::Create` uses the standard library
types `std::vector` and `std::unique_ptr` for one of its arguments,
which can't be used by an embedder built with a different standard
library.

This change adds a new version of the `Create()` function to
`v8::CppHeap` which doesn't accept the `std::vector`, since our
embedding use cases don't make use of it anyway. A `Destroy()`
function is added as well, to avoid issues with the embedder's
own implementation of the `delete` operator.