Enable clipping logic when `bb-simple-overflow-clip: auto`
This change will result in the overflow clip property being set in the
`ObjectPaintProperties` for the `LayoutObject`.  This is normally done
when the `overflow` CSS property is set, but now it will also be done
when the `bb-simple-overflow-clip` CSS property is set.

The simple implementation of the `clip_rect` calculation is taken from
the implementation of `LayoutBox::OverflowClipRect`, so it should match
closely what happens when the `overflow` CSS property is set.

Context: https://jira.prod.bloomberg.com/browse/ENG2BBKIT-7217