Hoist clip property out of paint chunks when `bb-simple-overflow-clip: auto`
Normally when clipping is implemented using the `overflow` CSS property,
the clip rect is stored as one of the chunk properties (along with
transforms and effects).  This results in separate chunks being created
for each element that clips using the `overflow` CSS property.

In the case of the TKS grid, each cell clips its content using the
`overflow` CSS property, which results in thousands of additional paint
chunks being created.

There are several places where having a large number of paint chunks can
cause complexity to increase and negatively impact performance:

* when collecting `PendingLayer`s
  https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/compositing/paint_artifact_compositor.cc;l=390;drc=3344b61f7c7f06cf96069751c3bd64d8ec3e3428

* matching old chunks to new chunks during raster invalidation:
  https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/paint/raster_invalidator.cc;l=191;drc=3344b61f7c7f06cf96069751c3bd64d8ec3e3428

These two operations in particular showed up highly when analyzing TKS
profiling data.

This commit attempts to avoid the creation of all these additional paint
chunks for the simple case of cell clipping.  We will use the custom CSS
property that was added previously: `bb-simple-overflow-clip`.

When this CSS property is set, we will hoist the clip portion from the
fragment's `ContentsProperties` out, so that it doesn't get used as part
of the chunk properties.  Effectively, the cells will be painted using
the same chunk (assuming no transforms or effects are applied on the
cell).

In order to implement the clip functionality, a new implementation of
`DisplayItem` will be created, called `SwitchToClipDisplayItem`.  The
clip portion from the fragment's `ContentsProperties` will be stored
inside this custom display item, and will be used when compositing the
paint chunks and display items to the `cc` layer.

Context: https://jira.prod.bloomberg.com/browse/ENG2BBKIT-7217