Restore suspension window only when we allow a trace
Background and Motivation
-------------------------
We double the suspension window whenever we suspend the GPU log tracer
so that we progressively wait longer for a lengthy spew. We also restore
the suspension window back to its original state over time so that we
are only ever aggressive about suspending traces during the spew.

Previously we were attempting to converge the suspension window towards
the original value for every incoming trace. This is a problem because
if we are dropping a large set of traces, the suspension window will
always equate to a value very close to the original value.

Change Description
------------------
With this change, we converge the suspension window towards the original
value only when we are not within the suspension window. This prevents
the dropped messages from reducing the suspension window.