Override activation window only when it's rooted
Prior to this changeset, we returned the `RenderWidgetHostView`'s
`aura::Window` as an override to the `DesktopNativeWidget`'s
`aura::Window` as the target for activation. This was done in order to
make sure the logical focus moves into the webview when the OS window
receives the keyboard focus.

The caller of this function indirectly requires the target window to be
rooted under `aura::WindowTreeHost`. This requirement is almost always
satisfied. In some rare cases, it seems that there is a disconnect
between the `RenderWidgetHostView` and the `WindowTreeHost`.

To combat the segmentation fault caused by this corner case, a check is
added to make sure the `RenderWidgetHostView` was rooted under a
`WindowTreeHost`. When this assumption is violated, we return a
`nullptr`, which signals the caller to revert to the default window it
would've picked as the target. This default window would be the
`DesktopNativeWidgetAura`. In this case, we iteratively print the
parents of the `RenderWidgetHostViewAura` until we hit a wall. This
will help us locate a (hopefully the only) detachment point between the
`RenderWidgetHostViewAura` and the `WindowTreeHost`.