Setup basic logging
This commit adds logging functionality to `feature/blpwtk2`.  This code
was previously in the `ext/log-message-handler` branch.  However, since
logging is so valuable to debug issues (especially when doing
migrations), let's add some basic logging in the `feature/blpwtk2`
branch as well.

Note that the more complex stuff like throttling and console logging is
still in the `ext/log-message-handler` branch.