[m104] Create a ThreadPoolInstance in renderer-main mode
This fixes a crash that happens when in the renderer-main mode and the
browser is out of process.  This can be tested by running
`blpwtk2_shell` with the `--renderer-main-mode-host` flag.

This mimics the upstream behavior here where the `ThreadPoolInstance` is
created before starting the renderer:
https://source.chromium.org/chromium/chromium/src/+/main:content/app/content_main_runner_impl.cc;l=531;drc=13dced01cce80364299ad7663479ad65069db73d

Context: https://jira.prod.bloomberg.com/browse/ENG2KITFC-1586