CppHeap: Common Initialization
When creating the CppHeap, a WrapperDescriptor is passed which configures which wrappers will be included in wrapper tracing. One of the values is an embedder id, and all wrappers which are intended to be traced need to have the same embedder id in their internal field.

Chromium initializes the CppHeap in the renderer process with the embedder id set to gin::kEmbedderBlink (1).

Chromium doesn't initialize the CppHeap in the browser process, as they don't create any wrappers there, but we do so we have explicit initialization in blpwtk2 for the browser process.

This commit ensures that the value specified for the embedder id matches that of the renderer process so that we can use a consistent value in our wrappers without having to know if we're running in the renderer or browser process.