Conditionally use default printer (#634)
Background and Motivation
-------------------------
Chromium can either be set to use the Windows default printer or to ask
the user for a printer selection. Traditionally our use case was to let
the embedder select a specific printer (not necessarily the Windows
default printer) and let Chromium use it without asking for a selection.
This was done by setting Chromium to always use the Windows default
printer and then by overriding the default printer inside the browser
process (not the actual Windows default printer).

We now want the ability for the embedder to unselect a printer so that
blpwtk2 will ask the user for a printer selection.

Change description
------------------
Prior to this commit, we did the following when a printer was specified
by the embedder:

  1. switch Chromium to use the default printer
  2. override the default printer selection

If the embedder specified a blank printer name, we still did both
operations above. This changeset makes it so that the switch to use
default printer only happens if the specified printer name is non-blank.
This will allow the embedder to specify a non-blank printer name and
have Chromium use it without prompting for a printer selection, or the
embedder can specify a blank printer name and Chromium will prompt for a
printer selection.