Move GN default args to .gn
This makes it so that whenever we change the default args for GN, we
don't need to re-run `gn.py`.  Internally, ninja will detect changes to
the `.gn` file, and update the build settings automatically.

This is for the change in the `blpwtk2/common` branch.  Will do the
remaining args in the `blpwtk2/feature/blpwtk2` branch later.