Combine hunspell suggestions and native spellcheck suggestions
Currently, if hunspell suggestions is available, we don't request
for native spellcheck suggestions, which is wrong.
Sometimes user have spellcheck enabled for multiple languages and those languages
could be using a mix of hunspell and native spellcheck. And the current implementation
will miss the suggestions from the native spellcheck

This commit mixes suggestions from hunspell and native spellcheck using
`spellcheck::FillSuggestions` which can remove the duplicates and order
the suggestions in a meaningful way.