Avoid crash when `child_process_launcher_` is null (#686)
This is neccessary due to the following upstream commit:
https://chromium.googlesource.com/chromium/src/+/d579bc29d0bcf765a8e657e04e43bcb74fc8864a

That commit moves the code to obtain the `ChildProcessTerminationInfo`
from `ProcessDied` to `GetChildTerminationInfo`.  However, it no longer
handles the case where `child_process_launcher_` is null.  This can be
the case in embedders that launch their own renderer processes, instead
of letting the browser manage the renderer processes.

This commit makes it handle the case where `child_process_launcher_` is
null, by simply returning the `ChildProcessTerminationInfo` with
`exit_code` set to 0 (matching the behavior prior to the upstream
change).

Context: https://jira.prod.bloomberg.com/browse/ENG2BBKIT-7832